move: 회원가입 타이틀과 입력폼 개별 컴포넌트로 분리
