+++ src/components/SignUp/index.tsx
@@ -1,0 +1,2 @@
+export { default as SignUpForm } from './SignUpForm';
+export { default as SignUpTitle } from './SignUpTitle';
